Updated plots and manuscript
I attach plots with preconditioner and without. It give a little more understanding of behaviour of each method, at least I think so. We should discuss that.
In terms of manuscript, I am sorry, it is quite rough draft, but I am struggling with what should be included in terms of mathematics and derivation of methods.